% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_fns.R
\name{add.m}
\alias{add.m}
\title{Modular Addition}
\usage{
add.m(a, b, m)
}
\arguments{
\item{a}{The first summand - integer}

\item{b}{The second summand - integer}

\item{m}{The modulo value - integer}
}
\description{
Adds two numbers and then returns the
remainder after dividing by the modulo.
}
\examples{
add.m(a=3 b= 9, n = 10)
}
