% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_fns.R
\name{build.ds}
\alias{build.ds}
\title{Construct a Difference Set Object}
\usage{
build.ds(ds, h)
}
\arguments{
\item{ds}{a vector of integers}

\item{h}{a group object (built via build.group function)}
}
\description{
Takes a vector of n integers and a group object
then it calls is.diff.set, if it passes, the function constructs
a difference set object for building a frame.
}
\examples{
build.ds(ds = as.integer(c(0,1,3), h=group)

}
