% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_fns.R
\name{build.group}
\alias{build.group}
\title{Build a Group Object}
\usage{
build.group(g)
}
\arguments{
\item{g}{a vector of integers}
}
\description{
Takes a vector of integers and then calls
is.group and determines if the given vector
is a group (currently fixed as integer cyclic)
Finally, it constructs a group object and returns it.
}
\examples{
build.group(g = as.integer(0:5))
}
