% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_fns.R
\name{count}
\alias{count}
\title{Count Occurences in a Vector}
\usage{
count(vect, n)
}
\arguments{
\item{vect}{a vector of values}

\item{n}{a number to be compared to the vector}
}
\description{
Counts the number of times the given number
n appears in the given vector vect.
}
\examples{
count(vect = 1:5, n = 4)
}
