% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frm_fns.R
\name{frame.ds}
\alias{frame.ds}
\title{Construct a Frame From a Difference Set}
\usage{
frame.ds(ds)
}
\arguments{
\item{ds}{a difference set object constructed via
build.ds function}
}
\description{
Builds the character table from the
given group and extracts the rows associated with the
given difference set.
}
\examples{
frame.ds(diff.set.object)

}
