% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frm_fns.R
\name{frame.stein}
\alias{frame.stein}
\title{Construct a Frame via Steiner Method}
\usage{
frame.stein(bibd, had)
}
\arguments{
\item{bibd}{The adjacency matrix for a (V,K,1)
bibd where the columns represent the vectors
and the rows represent the blocks.}

\item{had}{A hadamard matrix (square matrix with
every entry equal to 1 modulus) where the number
of rows and columns are equal to the BIBD value
value for K - the number of blocks that contain
a given vector.}
}
\description{
Builds a Frame by embedding the
rows of the given hadamard matrix into the
adjacency matrix of the given BIBD(V,K,1).
}
\examples{
frame.stein(bibd,had)

}
