% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stein_fns.R
\name{is.bibd}
\alias{is.bibd}
\title{Check if a matrix is a BIBD}
\usage{
is.bibd(bibd)
}
\arguments{
\item{bibd}{a {0,1}-valued integer matrix.}
}
\description{
Takes a matrix and determines if it meets
the requirements for the adjacency matrix of a BIBD(V,K,1).
This means that it has V columns, B rows, and only 1,0 values.
Each column has a 1 in exactly K rows and any two columns have
a 1 in exactly 1 shared row.
}
\examples{
is.bibd(bibd = matrix)
}
