% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_fns.R
\name{is.diff.set}
\alias{is.diff.set}
\title{Check if a set is a Difference Set for a Group}
\usage{
is.diff.set(ds, h)
}
\arguments{
\item{ds}{a vector of integers}

\item{h}{a group object (built via build.group function)}
}
\description{
Takes a vector of n integers and determines if
it is a difference set for the given group object
}
\examples{
is.diff.set(ds = as.integer(c(0,1,3), h=group)
}
