% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_fns.R
\name{is.group}
\alias{is.group}
\title{Check if a set is a Group}
\usage{
is.group(g)
}
\arguments{
\item{g}{a vector of integers}
}
\description{
Takes a vector of n integers and determines if
it is a cyclic integer group of order n (or scalar - integer
- multiple).
}
\examples{
is.group(g = as.integer(0:5))
}
