% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stein_fns.R
\name{is.hadamard}
\alias{is.hadamard}
\title{Check if a matrix is a hadamard}
\usage{
is.hadamard(hmat)
}
\arguments{
\item{hmat}{A square matrix.}
}
\description{
Takes a matrix and determines if it is a
hadamard matrix (possibly complex, square matrix with
entries all equal 1 modulus)
}
\examples{
is.hadamard(hmat = matrix)
}
