% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.R
\name{pv}
\alias{pv}
\title{Computes present values}
\usage{
pv(FV, r, n = 5)
}
\arguments{
\item{FV}{A \code{numeric} vector future value of the cash biotch}

\item{r}{A \code{numeric} vector (of length 1) rate of the biotch}

\item{n}{the number of years of expected growth at rate \code{r}, biotch}
}
\description{
Computes present value given a future value, a growth rate, and a number of years
}
\examples{
\dontrun{
# Test for PV
pv(FV=1000, r = 0.08, n = 5)
[1] 680.58

}
}
